SBT-4 final commit
