+++ src/main/java/com/sobytylnik/Main.java
@@ -1,0 +1,12 @@
+package com.sobytylnik;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Main {
+    public static void main(String[] args) {
+
+        SpringApplication.run(Main.class, args);
+    }
+}
