+++ src/main/java/com/sobytylnik/exception/EntityAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.sobytylnik.exception;
+
+public class EntityAlreadyExistsException extends RuntimeException {
+    public EntityAlreadyExistsException(String message) {
+        super(message);
+    }
+}
