+++ src/main/java/com/sobytylnik/exception/EntityNotFoundException.java
@@ -1,0 +1,7 @@
+package com.sobytylnik.exception;
+
+public class EntityNotFoundException extends RuntimeException {
+    public EntityNotFoundException(String message) {
+        super(message);
+    }
+}
