--- src/main/java/com/sobytylnik/hello/ExceptionListPackage/EntityAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package com.sobytylnik.hello.ExceptionListPackage;
-
-public class EntityAlreadyExistsException extends RuntimeException{
-    public EntityAlreadyExistsException(String message){
-        super(message);
-    }
-}
