--- src/main/java/com/sobytylnik/hello/ExceptionListPackage/EntityNotFoundException.java
@@ -1,7 +1,0 @@
-package com.sobytylnik.hello.ExceptionListPackage;
-
-public class EntityNotFoundException extends RuntimeException{
-    public EntityNotFoundException(String message){
-        super(message);
-    }
-}
