--- src/main/java/com/sobytylnik/hello/IndexController.java
@@ -1,14 +1,0 @@
-package com.sobytylnik.hello;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class IndexController {
-    private static final String HELLO = "Hello";
-
-    @RequestMapping("/")
-    public String index(){
-        return HELLO;
-    }
-}
