--- src/main/java/com/sobytylnik/hello/Main.java
@@ -1,12 +1,0 @@
-package com.sobytylnik.hello;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Main {
-    public static void main(String[] args) {
-
-        SpringApplication.run(Main.class, args);
-    }
-}
