--- src/main/java/com/sobytylnik/hello/ProfileRepository.java
@@ -1,16 +1,0 @@
-package com.sobytylnik.hello;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface ProfileRepository {
-
-    Optional<Profile> findById(Long id);
-
-    /**
-     * @return the profile with generated id.
-     */
-    Profile save(Profile profile);
-
-    void merge(Profile profile);
-}
